refactor: move PictureSelector to the ui folder and update import path
